requirements.txt
feedparser
gitpython
# heroku3
psycopg2-binary
pyrogram
python-dotenv
Pyromod==1.5
speedtest-cli
sqlalchemy==1.3.23
TgCrypto
